Moved Controller files to the newly created Controller folder. Created a Player Controller. Updated includes. Changed AI Melee characters health from 400 to 100.
