+++ Source/Sunrise/Controller/SunrisePlayerController.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/PlayerController.h"
+#include "SunrisePlayerController.generated.h"
+
+/**
+ * Controls the player.
+ */
+UCLASS()
+class SUNRISE_API ASunrisePlayerController : public APlayerController
+{
+	GENERATED_BODY()
+	
+};
